feat(einvoicing): Refactor building factur_x
