+++ app/services/e_invoice/factur_x/create/trade_delivery.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+module EInvoice
+  module FacturX
+    module Create
+      class TradeDelivery < Builder
+        def call
+          xml.comment "Applicable Header Trade Delivery"
+          xml["ram"].ApplicableHeaderTradeDelivery do
+            xml["ram"].ActualDeliverySupplyChainEvent do
+              xml["ram"].OccurrenceDateTime do
+                xml["udt"].DateTimeString '20250527', format: 102
+              end
+            end
+          end
+        end
+      end
+    end
+  end
+end
