+++ app/services/e_invoice/factur_x/create/trade_transaction.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module EInvoice
+  module FacturX
+    module Create
+      class TradeTransaction < Builder
+        def call(&block)
+          xml.comment "Supply Chain Trade Transaction"
+          xml['rsm'].SupplyChainTradeTransaction do
+            yield xml
+          end
+        end
+      end
+    end
+  end
+end
